third_party/WebKit/LayoutTests/svg/as-image/svg-canvas-link-not-colored-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
PASS window.colorChannels[0] is 0
PASS window.colorChannels[1] is 0
PASS window.colorChannels[2] is 0
PASS window.colorChannels[0] is 0
PASS window.colorChannels[1] is 128
PASS window.colorChannels[2] is 0
